data/toy_data_creation_repo/requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.7.0
PyYAML>=6.0
pandas>=1.3.0
